LayoutTests/fast/dom/HTMLProgressElement/progress-element-markup.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../../resources/dump-as-markup.js"></script>
</head>
<body>
  <div id="target"><progress style="-webkit-appearance: none;" value=70 max=100></progress></div>
<script>
Markup.dump(document.getElementById("target"), "A progress element should have a nested shadow box with a width specified");
</script>
</body>
</html>
