LayoutTests/fast/dom/HTMLProgressElement/progress-element-markup-expected.txt


A progress element should have a nested shadow box with a width specified:
| <progress>
|   max="100"
|   style="-webkit-appearance: none;"
|   value="70"
|   <shadow:root>
|     <div>
|       shadow:pseudoId="-webkit-progress-inner-element"
|       <div>
|         shadow:pseudoId="-webkit-progress-bar"
|         <div>
|           style="width: 70%; "
|           shadow:pseudoId="-webkit-progress-value"
